+++ CHANGELOG.md
@@ -8,13 +8,14 @@ NOTE: all lowercase and no ending punctuation marks
 -->
 
 ## Next release v1.0
+- [PB] - [#21](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/21) app(screen): set status for specific audience
+- [LE] - [#16](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/16) app(screen): new message screen
 
 - [CY] - [#18](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/18) app(screen): chat box screen
 - [PB] - [#24](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/24) app(feature): bubble see all modal screen
 - [PB] - [#25](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/25) app(feature): bubble screen
 - [PB] - [#19](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/19) app(screen): bubble screen
 - [PB] - [#13](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/13) app(screen): settings screen
-- [LE] - [#16](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/16) app(screen): new message screen
 - [CY] - [#11](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/11) app(screen): messages screen
 - [CY] - [#9](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/9) app(feature): home bottom tabs navigation
 - [CY] - [#7](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/7) app(feature): add dependencies to enable expo to publish the app online